+++ ui/src/helpers/avatar-size.js
@@ -1,3 +1,0 @@
-export default function () {
-  return this.$q.screen.lt.sm ? '145px' : '188px'
-}