+++ ui/src/helpers/status.js
@@ -2,17 +2,17 @@ const setStatusLabel = (status, value) => {
   const statusList = {
     active: { label: 'Ativo', color: 'primary' },
     inactive: { label: 'Inativo', color: 'negative' },
-    in_queue: { label: 'Em aprovação', color: 'warning' },
+    inQueue: { label: 'Em aprovação', color: 'warning' },
     reproved: { label: 'Reprovado', color: 'negative' },
     blocked: { label: 'Bloqueado', color: 'primary' },
     sold: { label: 'Vendido', color: 'positive' },
-    in_analysis: { label: 'Em análise', color: 'warning' },
-    available_for_sale: { label: 'Disponível para venda', color: 'positive' },
-    passed_on: { label: 'Repassado', color: 'primary-color' },
+    inAnalysis: { label: 'Em análise', color: 'warning' },
+    availableForSale: { label: 'Disponível para venda', color: 'positive' },
+    passedOn: { label: 'Repassado', color: 'primary-color' },
     unavailable: { label: 'Indisponível', color: 'negative' },
-    in_commercialization: { label: 'Em comercialização', color: 'positive' },
-    ready_to_commercializate: { label: 'Pronto para comercialização', color: 'primary' },
-    in_transfer: { label: 'Em repasse', color: 'positive' },
+    inCommercialization: { label: 'Em comercialização', color: 'positive' },
+    readyToCommercializate: { label: 'Pronto para comercialização', color: 'primary' },
+    inTransfer: { label: 'Em repasse', color: 'positive' },
     holding: { label: 'Aguardando', color: 'warning' }
   }