+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.x.x - 2021-04-27
+
+### Added
+- Added helper filter-object
+- Added helper status
+### Changed
+- Changed QasProfile layout and structure
+
 ## 1.x.x - 2021-04-22
 
 ### Added