+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added mixin `screen`
 - Added QasTextTruncate component in `ui/src/component`
+- Added `QasNestedFields`
 
 ### Changed
 - Added method in historyHandler for destroy routes by names list